[PL] Rename header file for uniform Dirichlet BC.
